backend/requirements.txt
asgiref==3.11.1
black==26.3.1
click==8.3.2
Django==6.0.4
django-cors-headers==4.9.0
django-environ==0.13.0
django-stubs==6.0.2
django-stubs-ext==6.0.2
djangorestframework==3.17.1
djangorestframework-stubs==3.16.9
mypy_extensions==1.1.0
packaging==26.1
pathspec==1.0.4
platformdirs==4.9.6
pytokens==0.4.1
sqlparse==0.5.5
types-PyYAML==6.0.12.20260408
typing_extensions==4.15.0
